requirements.txt
numpy>=1.22
networkx>=3.0
scikit-learn>=1.0
polars>=0.20
pandas>=1.5
matplotlib>=3.5
torch>=2.11
torchcde-0.2.5